+++ packages/ab-experiments/tsconfig.json
@@ -6,6 +6,9 @@
       "@titicaca/fetcher": ["../fetcher/src"],
       "@titicaca/fetcher/common": ["../fetcher/src/app-directory/common"],
       "@titicaca/react-contexts": ["../react-contexts/src"],
+      "@titicaca/react-contexts/common": [
+        "../react-contexts/src/app-directory/common"
+      ],
       "@titicaca/react-contexts/client": [
         "../react-contexts/src/app-directory/client"
       ],
