+++ packages/ad-banners/tsconfig.json
@@ -9,6 +9,9 @@
       "@titicaca/intersection-observer": ["../intersection-observer/src"],
       "@titicaca/router": ["../router/src"],
       "@titicaca/react-contexts": ["../react-contexts/src"],
+      "@titicaca/react-contexts/common": [
+        "../react-contexts/src/app-directory/common"
+      ],
       "@titicaca/react-contexts/client": [
         "../react-contexts/src/app-directory/client"
       ],
