+++ packages/carousel/tsconfig.json
@@ -7,6 +7,9 @@
       "@titicaca/icons": ["../icons/src"],
       "@titicaca/intersection-observer": ["../intersection-observer/src"],
       "@titicaca/react-contexts": ["../react-contexts/src"],
+      "@titicaca/react-contexts/common": [
+        "../react-contexts/src/app-directory/common"
+      ],
       "@titicaca/react-contexts/client": [
         "../react-contexts/src/app-directory/client"
       ],
