+++ packages/react-contexts/package.json
@@ -24,6 +24,9 @@
       "client": [
         "lib/app-directory/client/index.d.ts"
       ],
+      "common": [
+        "lib/app-directory/common/index.d.ts"
+      ],
       "middleware": [
         "lib/middleware.d.ts"
       ]
@@ -44,6 +47,11 @@
       "import": "./lib/app-directory/client/index.js",
       "require": "./lib/app-directory/client/index.js",
       "types": "./lib/app-directory/client/index.d.ts"
+    },
+    "./common": {
+      "import": "./lib/app-directory/common/index.js",
+      "require": "./lib/app-directory/common/index.js",
+      "types": "./lib/app-directory/common/index.d.ts"
     },
     "./middleware": {
       "import": "./lib/middleware.js",
