+++ packages/react-contexts/src/session-context/app-directory/server.ts
@@ -1,2 +1,1 @@
 export * from './get-session-provider-props'
-export { SessionContextProviderProps } from '../types'
