+++ packages/react-contexts/src/session-context/types.ts
@@ -20,5 +20,24 @@ export interface InAppSessionContextProviderProps {
 }
 
 export interface User {
+  name: string
+  provider: Provider
+  country: string
+  lang: string
+  unregister: boolean | null
+  photo: string
+  mileage: Mileage
   uid: string
+}
+
+type Provider = 'TRIPLE' | 'NAVER' | 'KAKAO' | 'FACEBOOK' | 'APPLE'
+
+interface Mileage {
+  badges: {
+    icon: {
+      imageUrl: string
+    }
+  }[]
+  level: number
+  point: number
 }
