+++ packages/react-contexts/src/session-context/user.ts
@@ -1,27 +1,6 @@
 import { createContext, useCallback, useContext, useState } from 'react'
 
-export interface User {
-  name: string
-  provider: Provider
-  country: string
-  lang: string
-  unregister: boolean | null
-  photo: string
-  mileage: Mileage
-  uid: string
-}
-
-type Provider = 'TRIPLE' | 'NAVER' | 'KAKAO' | 'FACEBOOK' | 'APPLE'
-
-interface Mileage {
-  badges: {
-    icon: {
-      imageUrl: string
-    }
-  }[]
-  level: number
-  point: number
-}
+import { User } from './types'
 
 export const GET_USER_REQUEST_URL = '/api/users/me'
