+++ packages/react-triple-client-interfaces/tsconfig.json
@@ -11,6 +11,9 @@
         "../view-utilities/src/app-directory/client"
       ],
       "@titicaca/react-contexts": ["../react-contexts/src"],
+      "@titicaca/react-contexts/common": [
+        "../react-contexts/src/app-directory/common"
+      ],
       "@titicaca/react-contexts/client": [
         "../react-contexts/src/app-directory/client"
       ],
