+++ packages/scrap-button/tsconfig.json
@@ -4,6 +4,9 @@
     "baseUrl": ".",
     "paths": {
       "@titicaca/react-contexts": ["../react-contexts/src"],
+      "@titicaca/react-contexts/common": [
+        "../react-contexts/src/app-directory/common"
+      ],
       "@titicaca/react-contexts/client": [
         "../react-contexts/src/app-directory/client"
       ],
