+++ packages/search/tsconfig.json
@@ -13,6 +13,9 @@
         "../view-utilities/src/app-directory/client"
       ],
       "@titicaca/react-contexts": ["../react-contexts/src"],
+      "@titicaca/react-contexts/common": [
+        "../react-contexts/src/app-directory/common"
+      ],
       "@titicaca/react-contexts/client": [
         "../react-contexts/src/app-directory/client"
       ],
