+++ packages/social-reviews/tsconfig.json
@@ -7,6 +7,9 @@
       "@titicaca/router": ["../router/src"],
       "@titicaca/i18n": ["../i18n/src"],
       "@titicaca/react-contexts": ["../react-contexts/src"],
+      "@titicaca/react-contexts/common": [
+        "../react-contexts/src/app-directory/common"
+      ],
       "@titicaca/react-contexts/client": [
         "../react-contexts/src/app-directory/client"
       ],
