+++ packages/triple-header/tsconfig.json
@@ -5,6 +5,9 @@
     "paths": {
       "@titicaca/core-elements": ["../core-elements/src"],
       "@titicaca/react-contexts": ["../react-contexts/src"],
+      "@titicaca/react-contexts/common": [
+        "../react-contexts/src/app-directory/common"
+      ],
       "@titicaca/react-contexts/client": [
         "../react-contexts/src/app-directory/client"
       ],
